+++ vclib-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oidc/helpers/AuthenticationResponsePreparationState.kt
@@ -1,6 +1,6 @@
 package at.asitplus.wallet.lib.oidc.helpers
 
-import at.asitplus.wallet.lib.data.dif.PresentationPreparationHelper
+import at.asitplus.wallet.lib.data.dif.PresentationPreparationState
 import at.asitplus.wallet.lib.oidc.AuthenticationRequestParameters
 import at.asitplus.wallet.lib.oidc.RelyingPartyMetadata
 import kotlinx.serialization.Serializable
@@ -14,5 +14,5 @@ data class AuthenticationResponsePreparationState(
     val audience: String,
     val responseModeParameters: ResponseModeParameters,
     val clientIdSchemeParameters: ClientIdSchemeParameters?,
-    val presentationPreparationHelper: PresentationPreparationHelper?,
+    val presentationPreparationState: PresentationPreparationState?,
 )