+++ vclib/src/commonMain/kotlin/at/asitplus/wallet/lib/data/dif/PresentationSubmissionValidator.kt
@@ -6,6 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 sealed interface PresentationSubmissionValidator {
     companion object {
+        @Throws(MissingInputDescriptorGroupException::class)
         fun createInstance(
             submissionRequirements: Collection<SubmissionRequirement>?,
             inputDescriptors: Collection<InputDescriptor>,
@@ -25,6 +26,33 @@ sealed interface PresentationSubmissionValidator {
             return KmmResult.success(verifier)
         }
     }
+
+    /**
+     * Checks, whether all submission requirements are satisfied
+     */
+    fun isSubmissionRequirementsSatisfied(
+        submittedInputDescriptorIds: Set<String>,
+    ): Boolean
+
+    /**
+     * Checks, whether submission requirements are satisfied, and also fails if there are unnecessary submissions
+     */
+    fun isValidSubmission(
+        submittedInputDescriptorIds: Set<String>,
+    ): Boolean {
+        return isSubmissionRequirementsSatisfied(submittedInputDescriptorIds) && findUnnecessaryInputDescriptorSubmissions(
+            submittedInputDescriptorIds
+        ).isEmpty()
+    }
+
+    fun findUnnecessaryInputDescriptorSubmissions(submittedInputDescriptorIds: Set<String>): Set<String> {
+        return submittedInputDescriptorIds.filter {
+            isSubmissionRequirementsSatisfied(
+                submittedInputDescriptorIds - it
+            )
+        }.toSet()
+    }
+
 
     @Serializable
     data class InputDescriptorSubmissionsValidator(
@@ -55,11 +83,7 @@ sealed interface PresentationSubmissionValidator {
         }
     }
 
-    fun isSubmissionRequirementsSatisfied(
-        submittedInputDescriptorIds: Set<String>,
-    ): Boolean
-
     class MissingInputDescriptorGroupException(inputDescriptor: InputDescriptor) : Exception(
-        "Input descriptor is missing field `group` and therefore is not eligible for use with submission requirements: $inputDescriptor"
+        "Input descriptor is missing field `group` and is therefore not eligible for use with submission requirements: $inputDescriptor"
     )
 }